requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
Flask==3.0.3
gunicorn==22.0.0
